glamor: Make sure we flush before Damage writes events to the client.
Long ago keithp had told me that BlockHandler would always be called
before events would go to the client, but we see no reason that that
would be the case today.  To ensure that X rendering is in the kernel
command stream before Damage goes to the client, flush at exactly that
point.

Signed-off-by: Eric Anholt <eric@anholt.net>