glamor: Rate-limit our BlockHandler flushing to every 5ms.
BlockHandler is called every time we might go to sleep waiting for a
client.  We've been flushing here to make sure that both ongoing
animation and any final rendering before all clients sleep appears on
the screen.  However, if clients are busy doing rendering that
frequently round-trips, this can be many more flushes than are
necessary.

XXX: Something is going wrong and x11perf doesn't seem to be flushing.