test: Add a helper script for testing glamor on a hardware driver.
I don't think we can integrate this with make check, and I'd rather
work on a modesetting-driver-as-headless-testing-environment series
for long term testing of glamor on hardware drivers.  Until then, this
should help driver developers get testing done of glamor on their
chip.

On my Skylake, this takes 12 minutes to run the test suite.

Signed-off-by: Eric Anholt <eric@anholt.net>