glamor: Avoid using a temporary in CopyArea by using BlitFramebuffer.
By default glBlitFramebuffer is undefined for overlapping copies, but
with a new GL extension (as yet undefined) the implementation can
advertise support for defined behavior on 1:1 overlap.  This can cut
the bandwidth for overlapping CopyArea (opaque window movement without
a compositor) in half.

XXX: Check the extension before taking this path.

Signed-off-by: Eric Anholt <eric@anholt.net>