simple-xinit: Reverse the arguments order to allow chaining servers.
One of the messes in building the test suite was that I needed shell
scripts to hide the '--' from one simple-xinit invocation from the
other.  If I put the server first and the client second, then we can
just ignore the second '--' argument as part of the second
simple-xinit invocation.

Signed-off-by: Eric Anholt <eric@anholt.net>