dri3: Use xcb's protocol headers instead of dri3proto.
This is an experiment as opposed to the "let's merge the xproto
repositories" plan -- can we use existing XCB headers to replace the
xproto repository usage entirely?

The downside is that we don't get an automatically-generated
DRI3_NAME.  Also, we're depending on the xcb *library* generated
headers, not a protocol-only header.

Signed-off-by: Eric Anholt <eric@anholt.net>